Modify module structure a bit
Iterator implementations now live in `iter`. That module as well as
`core` are now public. That was, the top-level namespace isn't that
polluted.
